pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "bs4>=0.0.2",
  "fastapi-mcp>=0.4.0",
  "fastapi[standard]==0.116.1",
  "google-generativeai>=0.8.5",
  "langchain-core>=0.3.72",
  "pytest>=8.4.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
